+++ packages/sdk/src/client/index.ts
@@ -1,6 +1,5 @@
-import { Psbt } from "bitcoinjs-lib";
+import { payments, Psbt } from "bitcoinjs-lib";
 import { Account } from "../account/types";
-import { TransactionFormat } from "../transaction/message";
 import { Network } from "../types";
 import { BitcoinUTXO, Output } from "../utxo";
 import { coinSelect } from "./coinselect";
@@ -9,12 +8,27 @@ import { getAddressType } from "../utils/address";
 import { AddressType } from "bitcoin-address-validation";
 import { encodeGlittrData } from "../utils/encode";
 import { fetchPOST } from "../utils/fetch";
+import { electrumFetchTxHex } from "../utils/electrum";
+import { OpReturnMessage } from "../transaction";
+
+export type CreateTxParams = {
+  address: string;
+  tx: OpReturnMessage;
+  outputs?: Output[];
+  utxos?: BitcoinUTXO[];
+};
 
 export type CreateBroadcastTxParams = {
   account: Account;
-  tx: TransactionFormat;
+  tx: OpReturnMessage;
   outputs?: Output[];
   utxos?: BitcoinUTXO[];
+};
+
+export type CreateAndBroadcastRawTxParams = {
+  account: Account;
+  inputs: BitcoinUTXO[];
+  outputs: Output[];
 };
 
 type GlittrSDKParams = {
@@ -32,55 +46,63 @@ export class GlittrSDK {
     this.network = network;
     this.glittrApi = glittrApi;
     this.electrumApi = electrumApi;
-
-    this.getUtxos = this.getUtxos.bind(this);
-    this.getTxHex = this.getTxHex.bind(this);
-    this.getGlittrAsset = this.getGlittrAsset.bind(this);
   }
 
-  private async getUtxos(address: string): Promise<BitcoinUTXO[]> {
-    // TODO glittr aware for txtype glittr
-    try {
-      const utxoFetch = await fetch(
-        `${this.electrumApi}/address/${address}/utxo`
-      );
-      const unconfirmedUtxos = (await utxoFetch.json()) ?? [];
-      const utxos = unconfirmedUtxos.filter(
-        (tx: BitcoinUTXO) => tx.status && tx.status.confirmed
-      );
+  async createTx({ address, tx, outputs, utxos }: CreateTxParams): Promise<Psbt> {
+    outputs = outputs ?? [];
+    const addressType = getAddressType(address);
 
-      return utxos;
-    } catch (e) {
-      throw new Error(`Error fetching UTXOS : ${e}`);
+    const embed = encodeGlittrData(JSON.stringify(tx));
+    outputs = outputs.concat({ script: embed, value: 0 });
+
+    const psbt = new Psbt({ network: getBitcoinNetwork(this.network) });
+    const coins = await coinSelect(
+      getBitcoinNetwork(this.network),
+      utxos ?? [],
+      outputs,
+      2,
+      address,
+      tx,
+      this.electrumApi,
+      this.glittrApi,
+      address
+    );
+
+    const _inputs = coins?.inputs ?? [];
+    for (const input of _inputs) {
+      switch (addressType) {
+        case AddressType.p2pkh:
+          psbt.addInput({
+            hash: input.hash,
+            index: input.index,
+            nonWitnessUtxo: input.nonWitnessUtxo,
+          });
+          break;
+        case AddressType.p2wpkh:
+          psbt.addInput({
+            hash: input.hash,
+            index: input.index,
+            witnessUtxo: input.witnessUtxo,
+          });
+          break;
+        default:
+          throw new Error(`Error Address Type not supported yet`);
+      }
     }
-  }
 
-  private async getTxHex(txId: string): Promise<string> {
-    try {
-      const txHexFetch = await fetch(`${this.electrumApi}/tx/${txId}/hex`);
-      const txHex = await txHexFetch.text();
-
-      return txHex;
-    } catch (e) {
-      throw new Error(`Error fetching TX Hex : ${e}`);
+    const _outputs = coins?.outputs ?? [];
+    for (const output of _outputs) {
+      if (output.address) {
+        psbt.addOutput({ address: output.address, value: output.value });
+      } else if (output.script) {
+        psbt.addOutput({ script: output.script, value: output.value });
+      }
     }
+
+    return psbt
   }
 
-  private async getGlittrAsset(txId: string, vout: number) {
-    try {
-      const assetFetch = await fetch(
-        `${this.glittrApi}/assets/${txId}/${vout}`
-      );
-      const asset = await assetFetch.text();
-
-      return JSON.stringify(asset);
-    } catch (e) {
-      throw new Error(`Error fetching Glittr Asset : ${e}`);
-    }
-  }
-
-  createTx() {}
-  broadcastTx() {}
+  // broadcastTx() {}
 
   async createAndBroadcastTx({
     account,
@@ -88,25 +110,31 @@ export class GlittrSDK {
     outputs,
     utxos,
   }: CreateBroadcastTxParams) {
-    outputs = outputs ?? []
+    outputs = outputs ?? [];
 
     const addressType = getAddressType(account.address);
 
     const embed = encodeGlittrData(JSON.stringify(tx));
-    outputs = outputs.concat({ script: embed, value: 0 });
+    outputs = [{ script: embed, value: 0 }, ...outputs];
 
     const psbt = new Psbt({ network: getBitcoinNetwork(this.network) });
     const coins = await coinSelect(
+      getBitcoinNetwork(this.network),
       utxos ?? [],
       outputs,
       2,
       account.address,
       tx,
-      this.getUtxos,
-      this.getTxHex,
-      this.getGlittrAsset,
+      this.electrumApi,
+      this.glittrApi,
       account.address
     );
+
+    // Hacky: If the embeded tx is different from the tx passed in, change the opreturn to the tx from coinselect
+    if (embed !== encodeGlittrData(JSON.stringify(coins?.tx))) {
+      const embedCoinTx = encodeGlittrData(JSON.stringify(coins?.tx));
+      outputs[0] = { script: embedCoinTx, value: 0 };
+    }
 
     const _inputs = coins?.inputs ?? [];
     for (const input of _inputs) {
@@ -148,13 +176,13 @@ export class GlittrSDK {
     const hex = psbt.extractTransaction(true).toHex();
 
     // Validate Glittr TX
-    const isValidGlittrTx = await fetchPOST(
-      `${this.glittrApi}/validate-tx`,
-      { "Content-Type": "application/json" },
-      hex
-    );
-    if (!isValidGlittrTx.is_valid)
-      throw new Error(`Glittr Error: TX Invalid ${isValidGlittrTx}`);
+    // const isValidGlittrTx = await fetchPOST(
+    //   `${this.glittrApi}/validate-tx`,
+    //   { "Content-Type": "application/json" },
+    //   hex
+    // );
+    // if (!isValidGlittrTx.is_valid)
+    //   throw new Error(`Glittr Error: TX Invalid ${isValidGlittrTx}`);
 
     // Broadcast TX
     const txId = await fetchPOST(
@@ -164,4 +192,73 @@ export class GlittrSDK {
     );
     return txId;
   }
+
+  async createAndBroadcastRawTx({
+    account,
+    inputs,
+    outputs
+  }: CreateAndBroadcastRawTxParams) {
+    const addressType = getAddressType(account.address);
+
+    if (inputs && inputs.length === 0) {
+      throw new Error("No inputs provided");
+    }
+
+    if (outputs && outputs.length === 0) {
+      throw new Error("No outputs provided");
+    }
+
+    const psbt = new Psbt({ network: getBitcoinNetwork(this.network) });
+
+    for (const input of inputs) {
+      switch (addressType) {
+        case AddressType.p2pkh:
+          const txHex = await electrumFetchTxHex(this.electrumApi, input.txid);
+          psbt.addInput({
+            hash: input.txid,
+            index: input.vout,
+            nonWitnessUtxo: Buffer.from(txHex, "hex"),
+          });
+          break;
+        case AddressType.p2wpkh:
+          const paymentOutput = payments.p2wpkh({
+            address: account.address,
+            network: getBitcoinNetwork(this.network)
+          }).output!;
+          psbt.addInput({
+            hash: input.txid,
+            index: input.vout,
+            witnessUtxo: {
+              script: paymentOutput,
+              value: input.value,
+            }
+          });
+          break;
+      }
+    }
+
+    for (const output of outputs) {
+      if (output.address) {
+        psbt.addOutput({ address: output.address, value: output.value });
+      } else if (output.script) {
+        psbt.addOutput({ script: output.script, value: output.value });
+      }
+    }
+
+    psbt.signAllInputs(account.keypair);
+    const isValidSignature = psbt.validateSignaturesOfAllInputs(validator);
+    if (!isValidSignature) {
+      throw new Error(`Error signature invalid`);
+    }
+    psbt.finalizeAllInputs();
+    const hex = psbt.extractTransaction(true).toHex();
+
+    const txId = await fetchPOST(
+      `${this.electrumApi}/tx`,
+      { "Content-Type": "application/json" },
+      hex
+    );
+    return txId;
+  }
 }
+