+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @glittr-sdk/sdk
 
+## 0.24.0
+
+### Minor Changes
+
+- Simple message
+
+## 0.23.0
+
+### Minor Changes
+
+- Simple message
+
+## 0.22.0
+
+### Minor Changes
+
+- Simple message
+
 ## 0.21.0
 
 ### Minor Changes